Change link directory to routes
