--- rent-a-car-ui/src/components/Car.tsx
@@ -1,3 +1,0 @@
-export const Car = () => {
-    return <div>CAR</div>
-}
