--- rent-a-car-ui/src/components/Rent.tsx
@@ -1,3 +1,0 @@
-export const Rent = () => {
-    return <div>RENT</div>
-}
