--- rent-a-car-ui/src/components/Settings.tsx
@@ -1,3 +1,0 @@
-export const Settings = () => {
-    return <div>SETTINGS</div>
-}
