+++ rent-a-car-ui/src/routes/Settings.tsx
@@ -1,0 +1,11 @@
+import Main from "../components/Main"
+import Sidebar from "../components/sidebar/Sidebar"
+
+export const Settings = () => {
+    return (
+        <div className="flex flex-1 h-full w-full ">
+            <Sidebar />
+            <Main />
+        </div>
+    )
+}
